# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: VrBook API
  version: 0.1.0
  description: |
    Direct-Booking Vacation Rental Platform — Phase 1 API.

    **Source of truth:** [`BookingApp_Proposal.md`](../BookingApp_Proposal.md).
    **Status:** v0.1 — every endpoint is stubbed; payload schemas are referenced but
    most are placeholders. Downstream agents replace stubs with real implementations
    per `proposal §20`.

    **Conventions** (proposal §6.1):
    - Base path: `/api/v1`. Major version always in URL.
    - Errors: RFC 7807 ProblemDetails. Validation: 400 with field-keyed `errors`.
    - Pagination: cursor for high-cardinality lists, offset for admin tables.
    - Idempotency: `Idempotency-Key` header on mutating endpoints; 24h dedupe.
    - Rate limits: 100/min anonymous, 600/min authenticated; `RateLimit-*` headers.
  contact:
    name: Solutions Architecture
servers:
  - url: http://localhost:8080/api/v1
    description: Local dev
  - url: https://api.vrbook.example.com/api/v1
    description: Production
security:
  - BearerAuth: []
...
paths:
...
  /properties:
    get:
      tags: [Catalog]
      summary: Public property search.
      security: []
      parameters:
        - { name: destination,  in: query, schema: { type: string } }
        - { name: checkin,      in: query, schema: { type: string, format: date } }
        - { name: checkout,     in: query, schema: { type: string, format: date } }
        - { name: guests,       in: query, schema: { type: integer, minimum: 1 } }
        - { name: min_price,    in: query, schema: { type: number } }
        - { name: max_price,    in: query, schema: { type: number } }
        - { name: amenities,    in: query, schema: { type: array, items: { type: string } }, style: form, explode: true }
        - { name: min_rating,   in: query, schema: { type: number, minimum: 1, maximum: 5 } }
        - { name: sort,         in: query, schema: { type: string } }
        - { name: cursor,       in: query, schema: { type: string } }
        - { name: limit,        in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
      responses:
        '200': { description: Search results. }
    post:
      tags: [Catalog]
      summary: Create a property (Owner).
      responses:
        '201': { description: Created. }
  /properties/{slug}:
    get:
      tags: [Catalog]
      summary: Public property detail.
      security: []
      parameters: [ { name: slug, in: path, required: true, schema: { type: string } } ]
      responses:
        '200': { description: Property. }
        '404': { $ref: '#/components/responses/Problem404' }
  /properties/{id}/availability:
    get:
      tags: [Catalog]
      summary: Day-by-day availability.
      security: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: from, in: query, required: true, schema: { type: string, format: date } }
        - { name: to,   in: query, required: true, schema: { type: string, format: date } }
      responses:
        '200': { description: Days. }
  /properties/{id}/images:
    post:
      tags: [Catalog]
      summary: Upload property image (multipart, Owner).
      parameters: [ { name: id, in: path, required: true, schema: { type: string, format: uuid } } ]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties: { file: { type: string, format: binary } }
      responses:
        '201': { description: Uploaded. }
  /properties/{id}/images/order:
    put:
      tags: [Catalog]
      summary: Reorder images (Owner).
      parameters: [ { name: id, in: path, required: true, schema: { type: string, format: uuid } } ]
      responses: { '204': { description: Reordered. } }
  /properties/{id}/images/{imageId}:
    delete:
      tags: [Catalog]
      summary: Delete an image (Owner).
      parameters:
        - { name: id,      in: path, required: true, schema: { type: string, format: uuid } }
        - { name: imageId, in: path, required: true, schema: { type: string, format: uuid } }
      responses: { '204': { description: Deleted. } }
...
  /feeds/{outboundToken}.ics:
    get:
      tags: ["Sync — Public Feed"]
      summary: Outbound iCal feed. Token-secured.
      security: []
      parameters: [ { name: outboundToken, in: path, required: true, schema: { type: string } } ]
      responses:
        '200':
          description: VCALENDAR.
          content: { text/calendar: { schema: { type: string } } }
        '404': { $ref: '#/components/responses/Problem404' }
...
  /admin/channel-feeds/{id}:
    put:
      tags: ["Sync — Admin"]
      parameters: [ { name: id, in: path, required: true, schema: { type: string, format: uuid } } ]
      responses: { '200': { description: Updated. } }
  /admin/channel-feeds/{id}/sync-now:
    post:
      tags: ["Sync — Admin"]
      summary: Trigger an immediate sync run.
      parameters: [ { name: id, in: path, required: true, schema: { type: string, format: uuid } } ]
      responses: { '202': { description: Queued. } }
...
  /admin/sync-conflicts/{id}/resolve:
    post:
      tags: ["Sync — Admin"]
      parameters: [ { name: id, in: path, required: true, schema: { type: string, format: uuid } } ]
      responses: { '200': { description: Resolved. } }

  # =========================== Messaging ==========================
...
  /threads/{id}/messages:
    get:
      tags: [Messaging]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: cursor, in: query, schema: { type: string } }
        - { name: limit,  in: query, schema: { type: integer, default: 50 } }
      responses: { '200': { description: Messages. } }
    post:
      tags: [Messaging]
      parameters: [ { name: id, in: path, required: true, schema: { type: string, format: uuid } } ]
      responses: { '201': { description: Sent. } }
  /threads/{id}/read:
    post:
      tags: [Messaging]
      parameters: [ { name: id, in: path, required: true, schema: { type: string, format: uuid } } ]
      responses: { '204': { description: Read receipt accepted. } }
  /threads/{id}/attachments:
    post:
      tags: [Messaging]
      parameters: [ { name: id, in: path, required: true, schema: { type: string, format: uuid } } ]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema: { type: object, properties: { file: { type: string, format: binary } } }
      responses: { '201': { description: Uploaded. } }
...
  /reviews/{id}/response:
    post:
      tags: [Reviews]
      summary: Owner replies to a review.
      parameters: [ { name: id, in: path, required: true, schema: { type: string, format: uuid } } ]
      responses: { '201': { description: Response posted. } }
  /properties/{id}/reviews:
    get:
      tags: [Reviews]
      summary: Public reviews list.
      security: []
      parameters: [ { name: id, in: path, required: true, schema: { type: string, format: uuid } } ]
      responses: { '200': { description: List. } }
...
  /admin/reviews/{id}/approve:
    post:
      tags: ["Reviews — Admin"]
      parameters: [ { name: id, in: path, required: true, schema: { type: string, format: uuid } } ]
      responses: { '200': { description: Approved. } }
  /admin/reviews/{id}/reject:
    post:
      tags: ["Reviews — Admin"]
      parameters: [ { name: id, in: path, required: true, schema: { type: string, format: uuid } } ]
      responses: { '200': { description: Rejected. } }

  # =========================== Admin / Reports ====================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: AD B2C bearer token.
  schemas:
...
    Problem:
      type: object
      properties:
        type:     { type: string, format: uri }
        title:    { type: string }
        status:   { type: integer, format: int32 }
        detail:   { type: string }
        instance: { type: string }
        traceId:  { type: string }
      additionalProperties: true
...
  responses:
...
    Problem404:
      description: Not found.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...